Time: 181 ms (42.86%), Space: 277.6 MB (75%) - LeetHub

+++ 3275-k-th-nearest-obstacle-queries/3275-k-th-nearest-obstacle-queries.java
@@ -1,20 +1,20 @@
 class Solution {
     class Pair {
         int[] point;
-        long dist;
-        public Pair(int[] point, long dist) {
+        int dist;
+        public Pair(int[] point, int dist) {
             this.point = point;
             this.dist = dist;
         }
     }
     public int[] resultsArray(int[][] queries, int k) {
-        Queue<Long> pq = new PriorityQueue<>(Collections.reverseOrder());
+        Queue<Integer> pq = new PriorityQueue<>(Collections.reverseOrder());
         int[] res = new int[queries.length];
         for(int i = 0; i<queries.length; i++) {
-            long dist = Math.abs(queries[i][0])+Math.abs(queries[i][1]);
+            int dist = Math.abs(queries[i][0])+Math.abs(queries[i][1]);
             pq.offer(dist);
             while(pq.size()>k) pq.poll();
-            res[i] = pq.size() == k ? pq.peek().intValue() : -1;
+            res[i] = pq.size() == k ? pq.peek() : -1;
         }
         return res;
     }
